requirements.txt
fastapi
uvicorn
python-multipart
ffmpeg-python
opencv-python
openai-whisper
llama-cpp-python
pydantic
jinja2
loguru
python-dotenv
